Fix dtype issue in compute_pos_weight function
修正内容:
- compute_pos_weight関数でマスクをfloat()に変換してからmean()を計算
- RuntimeError: mean() could not infer output dtype エラーを解消

動作確認:
- 2エポックの学習が正常に完了
- 損失が17.30 → 15.24に改善
- セグメンテーション損失が安定化（4.33 → 1.74）

+++ src/losses/segmentation.py
@@ -122,7 +122,8 @@ def compute_pos_weight(masks: torch.Tensor, min_ratio: float = 0.01) -> torch.Te
     Returns:
         pos_weight tensor
     """
-    pos_frac = masks.mean().clamp(min=min_ratio).item()
+    # floatに変換してから計算
+    pos_frac = masks.float().mean().clamp(min=min_ratio).item()
     neg_frac = 1.0 - pos_frac
     pos_weight = torch.tensor([neg_frac / pos_frac], device=masks.device)
     return pos_weight
